JOE 1.0.5
- Rectangle/Column mode.  I think I've made a very good implementation of 
this stuff.  It's really neat, I could have used this earlier.  When this 
mode is selected, the normal block commands (including filter block through 
unix command) and insert file work on rectangles.  Also if overtype mode is 
set, these commands overtype instead.  Also when rectangle mode is set, 
indent and unindent will work with the insertion/deletion point as the 
left side of the rectangle instead of the beginning of the lines.

- Can set tab width

- Can set indent/unindent step column-width and fill character

- ^K ;  now does a tags file search

- if there's only one help screen defined, then ^KH immediatly turns on help
instead of giving a menu

- same deal with TAB completion- if only one item matches, it does a simple
completion

- unnamed files get the name first given to save file ^K D

- joe is now in a subdirectory in the tar archive.  Also as you can see,
I'm including the version number along with the name

- the problems with the optimizer on Xenix and with back-tab on SCO UNIX are
fixed.  You still have to use 'ttysv' with SCO UNIX- I need to get access to
a SCO UNIX machine to make a tty interface for it.  I think the posix tty 
interface is probably pretty close.

- string searches were crashing when they got to the end of the file.  this 
is fixed now.